Add Child Assent
Added a cue to ask if child wants to play the game.
The assent occurs after the child says their name
and age and before instructions begin. If the child
presses the thumbs up, they proceed with the
instructions. Otherwise, the app goes to a thank you
page that allows the child to reload the page to
begin playing again.