Separate into 5 conditions
To counterbalance for the tree and have different prompts,
there will be 5 conditions instead of just "increasing" and
"constant". This is the combination of having the tall tree
on the left or right and whether the subject should expect
to bring the egg to the top of the tree on their first try combined
with a baseline.